requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.24.0
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Django==5.1.5
django-bootstrap5==24.3
django-cors-headers==4.6.0
django-environ==0.12.0
django-storages==1.14.4
idna==3.10
isodate==0.7.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.44.0
whitenoise==6.8.2
